Work around for invalid AppType on legacy gsm devices
Passion is reporting AppType as UNKNOWN, in 4.1 it seems to have fallenback on
type SIM but in 4.2 some restructuring prevents this. So to workaround force to type SIM

implemented as oldRilFeature apptypesim

Only use if you see in logcat.

D/RIL_UiccCardApplication(  567): Creating UiccApp: {APPTYPE_UNKNOWN,APPSTATE_DETECTED}
E/RIL_UiccCard(  567): App index 0 is invalid since it's not APPTYPE_SIM and not APPTYPE_USIM

Change-Id: I7c13ef8f2a30f5e00116e16691f44aef5a1401af